requirements.txt
azure-eventhub==5.11.5
azure-identity
azure-monitor-ingestion
